Update Listing Detail Header to new design
- render newsroom logo, if set in charter, in the header
- update layout per design
- Update Storyshots
